Refactoring of relational data unit implementation
