+++ commons-app/src/main/resources/osgiModuleFacadeConfig.properties
@@ -11,10 +11,10 @@ eu.unifiedviews.dataunit.relational;uses:="eu.unifiedviews.dataunit";version="${
 eu.unifiedviews.helpers.dataunit.copyhelper;uses:="eu.unifiedviews.dataunit.rdf,org.openrdf.query,org.openrdf.repository,org.openrdf.model,eu.unifiedviews.helpers.dataunit.rdfhelper,eu.unifiedviews.dataunit,org.slf4j,org.openrdf";version="${uv-dataunit-helpers.version}"
 eu.unifiedviews.helpers.dataunit.dataset;uses:="org.openrdf.query,org.openrdf.model,org.openrdf.query.impl";version="${uv-dataunit-helpers.version}"
 eu.unifiedviews.helpers.dataunit.fileshelper;uses:="eu.unifiedviews.dataunit.files,eu.unifiedviews.dataunit";version="${uv-dataunit-helpers.version}"
+eu.unifiedviews.helpers.dataunit.relationalhelper;uses:="eu.unifiedviews.dataunit.files,eu.unifiedviews.dataunit";version="${uv-dataunit-helpers.version}"
 eu.unifiedviews.helpers.dataunit.maphelper;uses:="eu.unifiedviews.dataunit,org.openrdf.repository,org.slf4j,org.openrdf,org.openrdf.query,eu.unifiedviews.helpers.dataunit.dataset,org.openrdf.model";version="${uv-dataunit-helpers.version}"
 eu.unifiedviews.helpers.dataunit.metadata;uses:="org.openrdf,org.openrdf.query.impl,org.openrdf.query,eu.unifiedviews.helpers.dataunit.dataset,org.openrdf.repository,eu.unifiedviews.dataunit,org.slf4j,org.openrdf.model";version="${uv-dataunit-helpers.version}"
 eu.unifiedviews.helpers.dataunit.rdfhelper;uses:="eu.unifiedviews.dataunit.rdf,org.openrdf.query,eu.unifiedviews.helpers.dataunit.dataset,eu.unifiedviews.dataunit,org.openrdf.model";version="${uv-dataunit-helpers.version}"
-eu.unifiedviews.helpers.dataunit.relationalhelper;uses:="eu.unifiedviews.dataunit.files,eu.unifiedviews.dataunit";version="${uv-dataunit-helpers.version}"
 eu.unifiedviews.helpers.dataunit.virtualgraphhelper;uses:="eu.unifiedviews.dataunit,org.openrdf,org.openrdf.query,eu.unifiedviews.helpers.dataunit.metadata,org.openrdf.repository,org.slf4j,org.openrdf.model";version="${uv-dataunit-helpers.version}"
 eu.unifiedviews.helpers.dataunit.virtualpathhelper;uses:="eu.unifiedviews.dataunit,org.openrdf,org.openrdf.query,eu.unifiedviews.helpers.dataunit.metadata,org.openrdf.repository,org.slf4j,org.openrdf.model";version="${uv-dataunit-helpers.version}"
 eu.unifiedviews.helpers.dataunit.resourcehelper;uses:="eu.unifiedviews.dataunit,eu.unifiedviews.helpers.dataunit.maphelper,org.slf4j";version="${uv-dataunit-helpers.version}"
