+++ dataunit-relational-impl/pom.xml
@@ -29,6 +29,18 @@
             <version>1.6.0-SNAPSHOT</version>
             <type>jar</type>
         </dependency>
+        
+        <!-- Internal relational database -->
+		<dependency>
+			<groupId>org.apache.commons</groupId>
+			<artifactId>commons-dbcp2</artifactId>
+			<version>2.0.1</version>
+		</dependency>
+		<dependency>
+			<groupId>com.h2database</groupId>
+			<artifactId>h2</artifactId>
+			<version>1.3.176</version>
+		</dependency>
 
         <dependency> <!-- JUnit dependency -->
             <groupId>junit</groupId>
