+++ dataunit-relational-impl/src/main/java/eu/unifiedviews/dataunit/relational/DataUnitDatabaseConnectionProvider.java
@@ -1,4 +1,4 @@
-package eu.unifiedviews.commons.dataunit.core;
+package eu.unifiedviews.dataunit.relational;
 
 import java.sql.Connection;
 import java.sql.SQLException;
