+++ dataunit-relational-impl/src/main/java/eu/unifiedviews/dataunit/relational/db/DatabaseWrapperConfigIF.java
@@ -1,4 +1,4 @@
-package eu.unifiedviews.commons.relational.db;
+package eu.unifiedviews.dataunit.relational.db;
 
 /**
  * Database wrapper configuration interface
