+++ dataunit-relational-impl/src/main/java/eu/unifiedviews/dataunit/relational/db/PooledConnectionProvider.java
@@ -1,4 +1,4 @@
-package eu.unifiedviews.commons.relational.db;
+package eu.unifiedviews.dataunit.relational.db;
 
 import java.sql.Connection;
 import java.sql.SQLException;
