+++ dataunit-relational-impl/src/main/java/eu/unifiedviews/dataunit/relational/repository/RelationalException.java
@@ -1,4 +1,4 @@
-package eu.unifiedviews.commons.relational.repository;
+package eu.unifiedviews.dataunit.relational.repository;
 
 public class RelationalException extends Exception {
