+++ dataunit-relational-impl/src/main/java/eu/unifiedviews/dataunit/relational/repository/RelationalRepositoryFactory.java
@@ -1,4 +1,4 @@
-package eu.unifiedviews.commons.relational.repository;
+package eu.unifiedviews.dataunit.relational.repository;
 
 import java.sql.Connection;
